requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-firestore
langchain
langchain-openai
python-dotenv
tiktoken
